Unison.Syntax.Name

Description

Syntax-related combinators for Name (to/from string types).

Documentation

fromText :: Text -> Maybe Name

Parse a name from a string literal.

Performs very minor validation (a name can't be empty, nor contain a # character [at least currently?]) but makes no attempt at rejecting bogus names like "foo...bar...baz".

fromTextEither :: Text -> Either Text Name

Parse a name from a string literal.

Performs very minor validation (a name can't be empty, nor contain a # character [at least currently?]) but makes no attempt at rejecting bogus names like "foo...bar...baz".

unsafeFromString :: String -> Name

Unsafely parse a name from a string literal. See unsafeFromText.

unsafeFromText :: HasCallStack => Text -> Name

Unsafely parse a name from a string literal.

Performs very minor validation (a name can't be empty, nor contain a # character [at least currently?]) but makes no attempt at rejecting bogus names like "foo...bar...baz".

unsafeFromVar :: Var v => v -> Name

Unsafely parse a name from a var, by first rendering the var as a string.

See unsafeFromText.

toString :: Name -> String

Convert a name to a string representation.

toText :: Name -> Text

Convert a name to a string representation.

toVar :: Var v => Name -> v

Convert a name to a string representation, then parse that as a var.

Orphan instances

IsString Name

Instance details

Methods

fromString :: String -> Name