Unison.Syntax.NameSegment

Description

Utilities related to the parsing and printing of name segments using the default syntax.

String conversions

toEscapedText :: NameSegment -> Text

Convert a name segment to escaped text, for display purposes.

toEscapedText (unsafeFromText ".~") = "`.~`"

toEscapedTextBuilder :: NameSegment -> Builder

parseText :: Text -> Either Text NameSegment

Parse text as a name segment.

parseText "foo" = Right (NameSegment "foo")
parseText ".~" = Left ...
parseText "`.~`" = Right (NameSegment ".~")

unsafeParseText :: Text -> NameSegment

Parse text as a name segment.

Name segment parsers

isSymboly :: NameSegment -> Bool

Name segment classifiers

data ParseErr

Constructors

ReservedOperator !Text

ReservedWord !Text

renderParseErr :: ParseErr -> Text

segmentP :: Monad m => ParsecT (Token ParseErr) [Char] m NameSegment

symbolyP :: ParsecT (Token Text) [Char] m NameSegment

A symboly name segment parser, which consists only of symboly characters.

A symboly name segment can optionally be escaped by surrounding it with backticks, which expands the list of allowed symbols to include these three: . ( )

Throws the parsed name segment as an error if it's unescaped and reserved, e.g. "=".

wordyP :: ParsecT (Token Text) [Char] m NameSegment

A wordy name segment parser, which consists only of wordy characters.

Throws the parsed name segment as an error if it's an unescaped keyword, e.g. "match".

Character classifiers

segmentStartChar :: Char -> Bool

symbolyIdChar :: Char -> Bool

wordyIdStartChar :: Char -> Bool

wordyIdChar :: Char -> Bool