Unison.Syntax.Parser

Documentation

class Annotated a where

Methods

ann :: a -> Ann

type Err v = ParseError Input (Error v)

data Error v

Constructors

SignatureNeedsAccompanyingBody (Token v)

DisallowedAbsoluteName (Token Name)

EmptyBlock (Token String)

UnknownTerm (Token (HashQualified Name)) (Set Referent)

UnknownType (Token (HashQualified Name)) (Set Reference)

UnknownId (Token (HashQualified Name)) (Set Referent) (Set Reference)

ExpectedBlockOpen String (Token Lexeme)

EmptyWatch Ann

UseInvalidPrefixSuffix (Either (Token Name) (Token Name)) (Maybe [Token Name])

UseEmpty (Token String)

DidntExpectExpression (Token Lexeme) (Maybe (Token Lexeme))

TypeDeclarationErrors [Error v Ann]

MissingTypeModifier (Token String) (Token v)
MissingTypeModifier (type|ability) name

TypeNotAllowed (Token (HashQualified Name))
A type was found in a position that requires a term

ResolutionFailures [ResolutionFailure Ann]

DuplicateTypeNames [(v, [Ann])]

DuplicateTermNames [(v, [Ann])]

PatternArityMismatch Int Int Ann
PatternArityMismatch expectedArity actualArity location

FloatPattern Ann

newtype Input

Constructors

Input

Fields

inputStream :: [Token Lexeme]

type P v m = ParsecT (Error v) Input (ReaderT (ParsingEnv m) m)

data ParsingEnv (m :: Type -> Type)

Constructors

ParsingEnv

Fields

uniqueNames :: UniqueName

uniqueTypeGuid :: Name -> m (Maybe Text)

Return a GUID to reuse for a unique type of the given name, if any.

This callback is called for every `unique type` declaration that does not explicitly specify a GUID.

The name (e.g. Foo in `unique type Foo`) is passed in, and if the function returns a Just, that GUID is used; otherwise, a random one is generated from $sel:uniqueNames:ParsingEnv.

names :: Names

maybeNamespace :: Maybe Name

localNamespacePrefixedTypesAndConstructors :: Names

newtype UniqueName

Constructors

UniqueName (Pos -> Int -> Maybe Text)

anyToken :: Ord v => P v m (Token Lexeme)

blank :: Ord v => P v m (Token NameSegment)

Parse a placeholder or typed hole

bytesToken :: Ord v => P v m (Token Bytes)

chainl1 :: Ord v => P v m a -> P v m (a -> a -> a) -> P v m a

Parse p 1+ times, combining with op

chainr1 :: Ord v => P v m a -> P v m (a -> a -> a) -> P v m a

chainl1Accum :: (Stream u, Ord s) => ParsecT s u m a -> ParsecT s u m (a -> a -> a) -> ParsecT s u m (a, [a -> a])

character :: Ord v => P v m (Token Char)

closeBlock :: Ord v => P v m (Token ())

Consume the end of a block

optionalCloseBlock :: Ord v => P v m (Token ())

With layout, blocks might “close” without an explicit outdent (e.g., not even a newline at the end of a Transclude). This allows those blocks to be closed by EOF.

doc :: Ord v => P v m (Token (UntitledSection (Tree (Token (ReferenceType, HashQualified Name)) [Token Lexeme])))

failCommitted :: Ord v => Error v -> P v m x

failureIf :: Ord v => P v m (P v m b) -> P v m a -> P v m b

If p would succeed, this fails uncommitted. Otherwise, failIfOk used to produce the output

hqInfixId :: Ord v => P v m (Token (HashQualified Name))

hqPrefixId :: Ord v => P v m (Token (HashQualified Name))

importSymbolyId :: Ord v => P v m (Token Name)

The + in: use Foo.bar + as a Name

importWordyId :: Ord v => P v m (Token Name)

Parse a wordyId as a Name, rejecting any hash

label :: (Ord v, Show a) => String -> P v m a -> P v m a

matchToken :: Ord v => Lexeme -> P v m (Token Lexeme)

Match a particular lexeme exactly, and consume it.

mkAnn :: (Annotated a, Annotated b) => a -> b -> Ann

numeric :: Ord v => P v m (Token String)

openBlock :: Ord v => P v m (Token String)

Consume a block opening and return the string that opens the block.

openBlockWith :: Ord v => String -> P v m (Token ())

peekAny :: Ord v => P v m (Token Lexeme)

positionalVar :: (Annotated a, Var v) => a -> v -> v

Gives this var an id based on its position - a useful trick to obtain a variable whose id won't match any other id in the file `positionalVar a Var.missingResult`

prefixDefinitionName :: Var v => P v m (Token v)

Parse a prefix identifier e.g. Foo or (+), discarding any hash

prefixTermName :: Var v => P v m (Token v)

Parse a prefix identifier e.g. Foo or (+), rejecting any hash This is useful for term declarations, where type signatures and term names should not have hashes.

queryToken :: Ord v => (Lexeme -> Maybe a) -> P v m (Token a)

Virtual pattern match on a lexeme.

reserved :: Ord v => String -> P v m (Token String)

Parse a reserved word

root :: Ord v => P v m a -> P v m a

rootFile :: Ord v => P v m a -> P v m a

run' :: (Monad m, Ord v) => P v m a -> String -> String -> ParsingEnv m -> m (Either (Err v) a)

run :: (Monad m, Ord v) => P v m a -> String -> ParsingEnv m -> m (Either (Err v) a)

semi :: Ord v => P v m (Token ())

Consume a virtual semicolon

seq :: Ord v => (Ann -> [a] -> a) -> P v m a -> P v m a

seq' :: Ord v => String -> (Ann -> [a] -> b) -> P v m a -> P v m b

sepBy :: Ord v => P v m a -> P v m b -> P v m [b]

sepBy1 :: Ord v => P v m a -> P v m b -> P v m [b]

string :: Ord v => P v m (Token Text)

symbolyDefinitionName :: Var v => P v m (Token v)

Parse a symboly ID like >>= or &&, discarding any hash

tok :: (Ann -> a -> b) -> Token a -> b

tokenToPair :: Token a -> (Ann, a)

tupleOrParenthesized :: Ord v => P v m a -> (Ann -> a) -> (a -> a -> a) -> P v m (Ann, a)

Parses a tuple of as, or a single parenthesized a

returns the result of combining elements with pair, alongside the annotation containing the full parenthesized expression.

uniqueBase32Namegen :: forall gen. DRG gen => gen -> UniqueName

uniqueName :: (Monad m, Var v) => Int -> P v m Text

wordyDefinitionName :: Var v => P v m (Token v)

Parse a wordy identifier e.g. Foo, discarding any hash

wordyPatternName :: Var v => P v m (Token v)

Orphan instances

Annotated a => Annotated (Pattern a)

Instance details

Methods

ann :: Pattern a -> Ann

(Annotated a, Annotated b) => Annotated (MatchCase a b)

Instance details

Methods

ann :: MatchCase a b -> Ann

Annotated a => Annotated (Term f v a)

Instance details

Methods

ann :: Term f v a -> Ann