Unison.Syntax.Parser.Doc.Data

Description

Haskell parallel to unison/base.Doc.

These types have two significant parameters: ident and code that are expected to be parameterized by some representation of identifiers and source code of the host language.

This is much more restricted than unison/base.Doc, but it covers everything we can parse from Haskell. The mismatch with Unison is a problem, as someone can create a Unison Doc with explicit constructors or function calls, have it rendered to a scratch file, and then we can’t parse it. Changing the types here to match Unison wouldn’t fix the issue. We have to modify the types and parser in concert (in both Haskell and Unison) to bring them in line.

Documentation

newtype UntitledSection a

Constructors

UntitledSection [a]

newtype Paragraph a

Constructors

Paragraph (NonEmpty a)

data List a

Constructors

BulletedList (NonEmpty (Column a))

NumberedList (NonEmpty (Word64, Column a))

data Column a

Constructors

Column (Paragraph a) (Maybe (List a))

data Top code leaf a

Constructors

Section (Paragraph leaf) [a]

Eval code

ExampleBlock code

CodeBlock String String

List' (List leaf)

Paragraph' (Paragraph leaf)

data EmbedLink a

This is a deviation from the Unison Doc data model – in Unison, Doc distinguishes between type and term links, but here Doc knows nothing about what namespaces may exist.

Constructors

EmbedLink a

newtype Transclude a

Constructors

Transclude a

newtype EmbedAnnotation ident a

Constructors

EmbedAnnotation (Either ident a)

data SourceElement ident a

Constructors

SourceElement (EmbedLink ident) [EmbedAnnotation ident a]

newtype EmbedSignatureLink a

Constructors

EmbedSignatureLink a

newtype Word

Constructors

Word String

newtype Join a

Constructors

Join (NonEmpty a)

newtype Group a

Constructors

Group (Join a)

data Leaf ident code a

Constructors

Link (EmbedLink ident)

NamedLink (Paragraph a) (Group a)
the Group always contains either a single Term/Type link or list of Transcludes & Words

Example code

Transclude' (Transclude code)

Bold (Paragraph a)

Italic (Paragraph a)

Strikethrough (Paragraph a)

Verbatim Word

Code Word

Source (NonEmpty (SourceElement ident (Transclude code)))

FoldedSource (NonEmpty (SourceElement ident (Transclude code)))

EvalInline code

Signature (NonEmpty (EmbedSignatureLink ident))

SignatureInline (EmbedSignatureLink ident)

Word' Word

Group' (Group a)