Unison.Syntax.Name

Description

Utilities related to the parsing and printing of names using the default syntax.

String conversions

parseText :: Text -> Maybe Name

Parse a name from a string literal.

parseTextEither :: Text -> Either Text Name

Parse a name from a string literal.

unsafeParseText :: HasCallStack => Text -> Name

Unsafely parse a name from a string literal.

toText :: Name -> Text

Convert a name to a string representation.

unsafeParseVar :: Var v => v -> Name

Unsafely parse a name from a var, by first rendering the var as a string.

See unsafeFromText.

parseVar :: Var v => v -> Maybe Name

Parse a name from a var, by first rendering the var as a string.

toVar :: Var v => Name -> v

Convert a name to a string representation, then parse that as a var.

Name parsers

nameP :: Monad m => ParsecT (Token ParseErr) [Char] m Name

A name parser.

relativeNameP :: forall m. Monad m => ParsecT (Token ParseErr) [Char] m Name

A relative name parser.

Name classifiers

isSymboly :: Name -> Bool